requirements.txt
streamlit
yfinance
pandas
plotly
pytest
lxml
sqlalchemy
psycopg[binary]
azure-monitor-opentelemetry